# openapi/legal-rag-openapi.yaml
openapi: 3.1.0
info:
  title: Legal Agentic RAG Internal Platform API
  version: 0.1.0
  description: Spec-first API for corpus ingestion, QA runtime, evaluation, gold annotation,
    and synthetic dataset generation.
servers:
- url: https://api.example.internal
...
paths:
...
  /v1/runs/{runId}:
    get:
      tags:
      - Runs
      summary: Get one inference run
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                    format: uuid
                  dataset_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  question_count:
                    type: integer
                  created_at:
                    type: string
                    format: date-time
                required:
                - run_id
                - dataset_id
                - status
                - question_count
                - created_at
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Inference run UUID
  /v1/runs/{runId}/questions/{questionId}:
    get:
      tags:
      - Runs
      summary: Get one prediction inside a run
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Inference run UUID
      - name: questionId
        in: path
        required: true
        schema:
          type: string
        description: Question ID
  /v1/runs/{runId}/export-submission:
    post:
      tags:
      - Runs
      summary: Export competition submission JSON
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  artifact_url:
                    type: string
                    format: uri
                  question_count:
                    type: integer
                required:
                - artifact_url
                - question_count
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                page_index_base:
                  type: integer
                  enum:
                  - 0
                  - 1
              required:
              - page_index_base
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Inference run UUID
...
  /v1/eval/runs/{evalRunId}:
    get:
      tags:
      - Evaluation
      summary: Get evaluation run status and metrics
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalRun'
      parameters:
      - name: evalRunId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Evaluation run UUID
...
  /v1/synth/jobs:
    post:
      tags:
      - Synthetic
      summary: Create a synthetic generation job
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyntheticJob'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyntheticJob'
  /v1/synth/jobs/{jobId}:
    get:
      tags:
      - Synthetic
      summary: Get synthetic generation job status
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyntheticJob'
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Synthetic job UUID
  /v1/synth/jobs/{jobId}/preview:
    post:
      tags:
      - Synthetic
      summary: Preview candidate questions before publication
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        candidate_id:
                          type: string
                          format: uuid
                        question:
                          type: string
                        answer_type:
                          type: string
                        source_pages:
                          type: array
                          items:
                            type: string
                      required:
                      - candidate_id
                      - question
                      - answer_type
                      - source_pages
                required:
                - items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
              required:
              - limit
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Synthetic job UUID
  /v1/synth/items/{candidateId}/approve:
    post:
      tags:
      - Synthetic
      summary: Approve, reject, or edit a synthetic candidate
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                - status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                decision:
                  type: string
                  enum:
                  - approve
                  - reject
                  - edit
                edited_question:
                  type:
                  - string
                  - 'null'
                edited_answer:
                  type:
                  - string
                  - number
                  - boolean
                  - 'null'
                  - array
                edited_source_pages:
                  type:
                  - array
                  - 'null'
                  items:
                    type: string
              required:
              - decision
      parameters:
      - name: candidateId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Synthetic candidate UUID
  /v1/synth/jobs/{jobId}/publish:
    post:
      tags:
      - Synthetic
      summary: Publish approved synthetic candidates as a dataset
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  dataset_id:
                    type: string
                    format: uuid
                  artifact_url:
                    type: string
                    format: uri
                required:
                - dataset_id
                - artifact_url
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                dataset_name:
                  type: string
                version:
                  type: string
              required:
              - dataset_name
              - version
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Synthetic job UUID
...
components:
  schemas:
...
    QueryResponse:
      title: QueryResponse
      description: Normalized answer plus sources and debug metadata.
      type: object
      additionalProperties: false
      properties:
        question_id:
          type: string
        answer:
          type:
          - string
          - number
          - boolean
          - 'null'
          - array
        answer_normalized:
          type:
          - string
          - 'null'
        answer_type:
          type: string
          enum:
          - boolean
          - number
          - date
          - name
          - names
          - free_text
        confidence:
          type: number
          minimum: 0
          maximum: 1
        route_name:
          type: string
        abstained:
          type: boolean
        sources:
          type: array
          items:
            $ref: '#/$defs/page_ref'
        telemetry:
          $ref: '#/$defs/telemetry'
        debug:
          type: object
          additionalProperties: false
          properties:
            entities:
              type: array
              items:
                type: object
            candidate_pages:
              type: array
              items:
                $ref: '#/$defs/page_ref'
            policy_version:
              type: string
            trace_url:
              type:
              - string
              - 'null'
          required:
          - policy_version
      required:
      - question_id
      - answer
      - answer_type
      - confidence
      - route_name
      - abstained
      - sources
      - telemetry
      $defs:
        page_ref:
          $schema: https://json-schema.org/draft/2020-12/schema
          title: PageRef
          description: Canonical PDF page reference used in evaluation and submission.
          type: object
          additionalProperties: false
          properties:
            project_id:
              type: string
              format: uuid
            document_id:
              type: string
              format: uuid
            pdf_id:
              type: string
              pattern: ^[A-Za-z0-9._-]+$
            page_num:
              type: integer
              minimum: 0
            page_index_base:
              type: integer
              enum:
              - 0
              - 1
            source_page_id:
              type: string
              pattern: ^[A-Za-z0-9._-]+_[0-9]+$
            used:
              type: boolean
            evidence_role:
              type: string
              enum:
              - primary
              - supporting
              - title
              - lineage
              - negative_check
            score:
              type: number
              minimum: 0
              maximum: 1
          required:
          - pdf_id
          - page_num
          - page_index_base
          - source_page_id
          - used
        telemetry:
          $schema: https://json-schema.org/draft/2020-12/schema
          title: Telemetry
          description: Per-answer telemetry captured for competition submission and
            internal evaluation.
          type: object
          additionalProperties: false
          properties:
            request_started_at:
              type: string
              format: date-time
            first_token_at:
              type:
              - string
              - 'null'
              format: date-time
            completed_at:
              type: string
              format: date-time
            ttft_ms:
              type: integer
              minimum: 0
            total_response_ms:
              type: integer
              minimum: 0
            time_per_output_token_ms:
              type:
              - number
              - 'null'
              minimum: 0
            input_tokens:
              type: integer
              minimum: 0
            output_tokens:
              type: integer
              minimum: 0
            model_name:
              type: string
            route_name:
              type: string
            judge_model_name:
              type:
              - string
              - 'null'
            search_profile:
              type: string
            telemetry_complete:
              type: boolean
            trace_id:
              type: string
          required:
          - request_started_at
          - completed_at
          - ttft_ms
          - total_response_ms
          - input_tokens
          - output_tokens
          - model_name
          - route_name
          - telemetry_complete
...
    EvalRun:
      title: EvalRun
      description: Evaluation run against a gold dataset.
      type: object
      additionalProperties: false
      properties:
        eval_run_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        gold_dataset_id:
          type: string
          format: uuid
        scoring_policy_version:
          type: string
        judge_policy_version:
          type: string
        status:
          type: string
          enum:
          - queued
          - running
          - completed
          - failed
        metrics:
          type: object
          additionalProperties: false
          properties:
            answer_score_mean:
              type: number
            grounding_score_mean:
              type: number
            telemetry_factor:
              type: number
            ttft_factor:
              type: number
            overall_score:
              type: number
            source_precision:
              type: number
            source_recall:
              type: number
            source_f_beta:
              type: number
            no_answer_precision:
              type: number
            no_answer_recall:
              type: number
            p50_ttft_ms:
              type: integer
            p95_ttft_ms:
              type: integer
          required:
          - answer_score_mean
          - grounding_score_mean
          - telemetry_factor
          - ttft_factor
          - overall_score
          - source_precision
          - source_recall
          - source_f_beta
      required:
      - eval_run_id
      - run_id
      - gold_dataset_id
      - scoring_policy_version
      - judge_policy_version
      - status
      - metrics
    SyntheticJob:
      title: SyntheticJob
      description: Synthetic dataset generation job.
      type: object
      additionalProperties: false
      properties:
        job_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - queued
          - running
          - review
          - published
          - failed
          - cancelled
        source_scope:
          type: object
          additionalProperties: false
          properties:
            document_ids:
              type: array
              items:
                type: string
            doc_types:
              type: array
              items:
                type: string
            date_range:
              type:
              - object
              - 'null'
              additionalProperties: false
              properties:
                from:
                  type: string
                  format: date
                to:
                  type: string
                  format: date
              required:
              - from
              - to
          required:
          - document_ids
          - doc_types
        generation_policy:
          type: object
          additionalProperties: false
          properties:
            target_count:
              type: integer
              minimum: 1
            answer_type_mix:
              type: object
              additionalProperties:
                type: number
                minimum: 0
                maximum: 1
            route_mix:
              type: object
              additionalProperties:
                type: number
                minimum: 0
                maximum: 1
            adversarial_ratio:
              type: number
              minimum: 0
              maximum: 1
            paraphrase_ratio:
              type: number
              minimum: 0
              maximum: 1
            require_human_review:
              type: boolean
          required:
          - target_count
          - answer_type_mix
          - route_mix
          - adversarial_ratio
          - paraphrase_ratio
          - require_human_review
      required:
      - job_id
      - project_id
      - status
      - source_scope
      - generation_policy
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []